[add] 4 lesson
